+++ project-system/src/main/java/com/software/system/entity/RoleDept.java
@@ -21,13 +21,13 @@ public class RoleDept implements Serializable {
 
     @ApiModelProperty("id")
     @TableId(value = "id", type = IdType.AUTO)
-    private Long id;
+    private Integer id;
 
     @ApiModelProperty("角色id")
     @TableField("role_id")
-    private Long roleId;
+    private Integer roleId;
 
     @ApiModelProperty("菜单id")
     @TableField("dept_id")
-    private Long deptId;
+    private Integer deptId;
 }
